Use new tabs everywhere
