+++ packages/app/src/ui/atoms/tabs/Tabs.tsx
@@ -2,40 +2,55 @@ import * as TabsPrimitive from '@radix-ui/react-tabs'
 import * as React from 'react'
 
 import { cn } from '@/ui/utils/style'
+import { VariantProps, cva } from 'class-variance-authority'
 
 const Tabs = TabsPrimitive.Root
 
+const tabsListVariants = cva('inline-flex w-full flex-wrap items-center justify-center rounded-sm bg-secondary', {
+  variants: {
+    size: {
+      l: 'typography-label-4 gap-2 p-1.5',
+      s: 'typography-label-6 gap-0.5 p-0.5',
+    },
+  },
+  defaultVariants: {
+    size: 'l',
+  },
+})
+
 const TabsList = React.forwardRef<
   React.ElementRef<typeof TabsPrimitive.List>,
-  React.ComponentPropsWithoutRef<typeof TabsPrimitive.List>
->(({ className, ...props }, ref) => (
-  <TabsPrimitive.List
-    ref={ref}
-    className={cn('inline-flex w-full items-center justify-center border-panel-border border-b', className)}
-    {...props}
-  />
+  React.ComponentPropsWithoutRef<typeof TabsPrimitive.List> & VariantProps<typeof tabsListVariants>
+>(({ className, size, ...props }, ref) => (
+  <TabsPrimitive.List ref={ref} className={cn(tabsListVariants({ size }), className)} {...props} />
 ))
 TabsList.displayName = TabsPrimitive.List.displayName
 
+const tabsTriggerVariants = cva(
+  cn(
+    'flex-1 whitespace-nowrap rounded-[6px] border border-transparent ease-in ',
+    'bg-secondary text-secondary transition-colors hover:bg-tertiary',
+    'data-[state=active]:border-primary data-[state=active]:bg-primary data-[state=active]:text-primary data-[state=active]:shadow-xs', // shows the blue bar for the active tab
+    'focus-visible:outline-none focus-visible:ring focus-visible:ring-reskin-primary-200 focus-visible:ring-offset-0',
+  ),
+  {
+    variants: {
+      size: {
+        l: 'p-2 md:px-4',
+        s: 'p-1.5 md:px-2',
+      },
+    },
+    defaultVariants: {
+      size: 'l',
+    },
+  },
+)
+
 const TabsTrigger = React.forwardRef<
   React.ElementRef<typeof TabsPrimitive.Trigger>,
-  React.ComponentPropsWithoutRef<typeof TabsPrimitive.Trigger> & {
-    indicatorClassName?: string
-  }
->(({ className, children, indicatorClassName, ...props }, ref) => (
-  <TabsPrimitive.Trigger
-    ref={ref}
-    className={cn(
-      'inline-flex w-full flex-grow flex-col items-center justify-center whitespace-nowrap px-0.5 font-semibold text-base opacity-50 transition-opacity duration-200 ease-in-out',
-      'data-[state=active]:opacity-100 hover:opacity-70 focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-offset-2',
-      '[&>*]:data-[state=active]:opacity-100', // shows the blue bar for the active tab
-      className,
-    )}
-    {...props}
-  >
-    {children}
-    <div className={cn('mt-2 h-1 w-[calc(100%-1.5rem)] rounded-t-lg bg-nav-primary opacity-0', indicatorClassName)} />
-  </TabsPrimitive.Trigger>
+  React.ComponentPropsWithoutRef<typeof TabsPrimitive.Trigger> & VariantProps<typeof tabsTriggerVariants>
+>(({ className, size, ...props }, ref) => (
+  <TabsPrimitive.Trigger ref={ref} className={cn(tabsTriggerVariants({ size }), className)} {...props} />
 ))
 TabsTrigger.displayName = TabsPrimitive.Trigger.displayName
 
@@ -46,7 +61,7 @@ const TabsContent = React.forwardRef<
   <TabsPrimitive.Content
     ref={ref}
     className={cn(
-      'ring-offset-background data-[state=active]:mt-3 focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2',
+      'data-[state=active]:mt-3 focus-visible:outline-none focus-visible:ring focus-visible:ring-reskin-primary-200 focus-visible:ring-offset-0',
       className,
     )}
     {...props}
